+++ packages/terra-consumer-layout/src/Layout.jsx
@@ -2,6 +2,8 @@ import React from 'react';
 import PropTypes from 'prop-types';
 import classNames from 'classnames/bind';
 import Nav from 'terra-consumer-nav';
+import ResponsiveElement from 'terra-responsive-element';
+import Overlay from 'terra-overlay';
 import { injectIntl, intlShape } from 'react-intl';
 import styles from './Layout.scss';
 
@@ -30,6 +32,7 @@ class Layout extends React.Component {
     this.state = {
       isMobileNavOpen: false,
     };
+
     this.toggleNav = this.toggleNav.bind(this);
   }
 
@@ -41,6 +44,15 @@ class Layout extends React.Component {
 
   render() {
     const { nav, helpItems, intl, ...customProps } = this.props;
+    const overlay = (
+      <Overlay
+        onRequestClose={this.toggleNav}
+        isOpen={this.state.isMobileNavOpen}
+        backgroundStyle="clear"
+        isRelativeToContainer
+      />
+    );
+
     return (
       <div>
         <div className={cx('skip-container')}>
@@ -55,6 +67,7 @@ class Layout extends React.Component {
             onRequestClose={this.toggleNav}
           />
           <main id="main-container" className={cx('main-container', this.state.isMobileNavOpen && 'nav-open')}>
+            <ResponsiveElement defaultElement={overlay} responsiveTo="window" medium={<div />} />
             <Nav.Burger className={cx('nav-burger')} handleClick={this.toggleNav} />
             <div className={cx('main-content')}>{this.props.children}</div>
             <Nav.Help className={cx('help-button')} helpNavs={helpItems} id="nav-help-button" />