+++ packages/terra-consumer-nav/tests/nightwatch/NavTestRoutes.jsx
@@ -8,6 +8,8 @@ import NavTests from './NavTests';
 import BurgerButtonTest from './BurgerButtonTest';
 import HelpButtonTest from './HelpButtonTest';
 import SimpleNav from './SimpleNav';
+import NavLogoTest from './NavLogoTest';
+import SimpleNavWithTextLogo from './SimpleNavWithTextLogo';
 
 const routes = (
   <div>
@@ -15,6 +17,8 @@ const routes = (
     <Route path="/tests/nav-tests/burger-button" component={BurgerButtonTest} />
     <Route path="/tests/nav-tests/help-button" component={HelpButtonTest} />
     <Route path="/tests/nav-tests/simple-nav" component={SimpleNav} />
+    <Route path="/tests/nav-tests/simple-nav-logo-text" component={SimpleNavWithTextLogo} />
+    <Route path="/tests/nav-tests/logo" component={NavLogoTest} />
   </div>
 );