+++ packages/terra-consumer-nav/CHANGELOG.md
@@ -10,10 +10,10 @@ ChangeLog
 ### Changed
 - Fix scroll side nav in mobile in Firefox and IE
 - Changed logo height to be max of 120 px for both mobile and desktop
-- Changed logo height to be max of 120 px for both mobile and desktop
 - Updated hover effects of nav items and profile links.
 - Stop using terra-button for nav close modal, profile button, and help button
 - Made Nav and Nav-Container width consistent in all places used.
+- Fixed modal content scrolling issue.
 
 ------------------