+++ packages/terra-consumer-nav/CHANGELOG.md
@@ -15,7 +15,7 @@ ChangeLog
 - Stop using terra-button for nav close modal, profile button, and help button
 - Made Nav and Nav-Container width consistent in all places used.
 - Fixed modal content scrolling issue.
-
+- Refactored NavLogo for flex width and added new styles to logo text
 ------------------
 
 # 0.1.0-BETA.5 - (September 06, 2017)