+++ packages/terra-consumer-nav/CHANGELOG.md
@@ -16,6 +16,7 @@ ChangeLog
 - Made Nav and Nav-Container width consistent in all places used.
 - Fixed modal content scrolling issue.
 - Refactored NavLogo for flex width and added new styles to logo text
+- Clean up custom properties
 ------------------
 
 # 0.1.0-BETA.5 - (September 06, 2017)